Refactor: move `Supabase` types to appropriate directory.
